feat: ISP를 위해 배치를 분리
